Date and count validation + More setup for Maven
Added some checks for dates and counts to see if they're valid before using them. They are omitted from the parsed JSON if invalid. If the input date is invalid, all map elements are disabled.
Fetch call for application layer is implemented. Until the fetch happens, it uses the dummy json data, which has been moved to the 'data' folder in 'presentation'.